# openapi.yml
openapi: 3.1.0
info:
  title: Toby Reminder API
  description: Apple Reminder App 웹 버전 REST API
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api/lists:
    get:
      tags:
        - ReminderList
      summary: 전체 리스트 조회
      description: 모든 리마인더 리스트를 displayOrder 순으로 조회한다.
      operationId: findAllLists
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReminderListResponse"

    post:
      tags:
        - ReminderList
      summary: 리스트 생성
      description: 새로운 리마인더 리스트를 생성한다.
      operationId: createList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderListRequest"
      responses:
        "201":
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderListResponse"

  /api/lists/{id}:
    get:
      tags:
        - ReminderList
      summary: 리스트 단건 조회
      description: ID로 리마인더 리스트를 조회한다.
      operationId: findListById
      parameters:
        - $ref: "#/components/parameters/ListId"
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderListResponse"
        "404":
          description: 리스트를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

    put:
      tags:
        - ReminderList
      summary: 리스트 수정
      description: 기존 리마인더 리스트를 수정한다.
      operationId: updateList
      parameters:
        - $ref: "#/components/parameters/ListId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderListRequest"
      responses:
        "200":
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderListResponse"
        "404":
          description: 리스트를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

    delete:
      tags:
        - ReminderList
      summary: 리스트 삭제
      description: 리마인더 리스트를 삭제한다.
      operationId: deleteList
      parameters:
        - $ref: "#/components/parameters/ListId"
      responses:
        "204":
          description: 삭제 성공
        "404":
          description: 리스트를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

components:
  parameters:
    ListId:
      name: id
      in: path
      required: true
      description: 리마인더 리스트 ID
      schema:
        type: integer
        format: int64

  schemas:
    ReminderListRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: 리스트 이름
          example: Shopping
        color:
          type: string
          description: "색상 코드 (#hex)"
          example: "#FF9500"
        icon:
          type: string
          description: 아이콘 이름
          example: cart
        displayOrder:
          type: integer
          description: 정렬 순서
          example: 1

    ReminderListResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 리스트 ID
          example: 1
        name:
          type: string
          description: 리스트 이름
          example: Shopping
        color:
          type: string
          description: "색상 코드 (#hex)"
          example: "#FF9500"
        icon:
          type: string
          description: 아이콘 이름
          example: cart
        displayOrder:
          type: integer
          description: 정렬 순서
          example: 1
        createdAt:
          type: string
          format: date-time
          description: 생성 시각
        updatedAt:
          type: string
          format: date-time
          description: 수정 시각

    ProblemDetail:
      type: object
      description: RFC 9457 Problem Detail
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri